функционал для работы с временными периодами перенесён в отдельный класс
